requirements.txt
streamlit
fastapi
sqlalchemy
uvicorn
openai
langchain
numpy
faiss-cpu
pytube
openai-whisper
ffmpeg-python
python-dotenv
